Abstract out authRoutes from server
