+++ server.js
@@ -3,9 +3,8 @@ const next = require("next")
 const session = require("express-session")
 const passport = require("passport")
 const Auth0Strategy = require("passport-auth0")
-const util = require("util")
-const url = require("url")
-const querystring = require("querystring")
+
+const authRoutes = require("./server/auth-routes")
 
 const {
   NODE_ENV,
@@ -75,54 +74,7 @@ app.prepare().then(() => {
   server.use(passport.initialize())
   server.use(passport.session())
 
-  // Perform the login, after login Auth0 will redirect to callback
-  server.get(
-    "/login",
-    passport.authenticate("auth0", {
-      scope: "openid email profile",
-    }),
-    (req, res) => {
-      res.redirect("/")
-    }
-  )
-
-  // Perform the final stage of authentication and redirect to previously requested URL or '/user'
-  server.get("/callback", (req, res, next) => {
-    passport.authenticate("auth0", (err, user) => {
-      if (err) return next(err)
-      if (!user) return res.redirect("/login")
-      req.logIn(user, err => {
-        if (err) return next(err)
-
-        const returnTo = req.session.returnTo
-        delete req.session.returnTo
-        res.redirect(returnTo || "/")
-      })
-    })(req, res, next)
-  })
-
-  // Perform session logout and redirect to homepage
-  server.get("/logout", (req, res) => {
-    req.logout()
-
-    let returnTo = req.protocol + "://" + req.hostname
-    const port = req.connection.localPort
-    if (port !== undefined && port !== 80 && port !== 443) {
-      returnTo += ":" + port
-    }
-
-    const logoutURL = new url.URL(
-      util.format("https://%s/v2/logout", AUTH0_DOMAIN)
-    )
-
-    const searchString = querystring.stringify({
-      client_id: AUTH0_CLIENT_ID,
-      returnTo,
-    })
-
-    logoutURL.search = searchString
-    res.redirect(logoutURL)
-  })
+  server.use(authRoutes)
 
   // server.get("/test", restrictAccessPage)
   server.get("/api/test", restrictAccessAPI)
